api/requirements.txt
azure-functions
flask
flask-cors
flask
flask-cors
pillow
tqdm
torch
transformers
scikit-learn
